Working on react native updates
