+++ App.tsx
@@ -2,31 +2,22 @@ import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { StyleSheet, Text, View } from 'react-native';
-import Homepage from './screens/Homepage';
-import Invoice from './screens/Invoice';
-import Profile from './screens/Profile';
-import Signup from './screens/auth/Signup';
-import NavigationBar from './components/Navbar';
+import { Homepage } from './screens/Homepage';
+import { Invoice } from './screens/Invoice';
+import { Profile } from './screens/Profile';
+import { NavigationBar } from './components/NavigationBar';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
 const Tab = createBottomTabNavigator();
 
 export default function App() {
-  return <Signup></Signup>;
-}
-
-const Stack = createNativeStackNavigator();
-
-const MyStack = () => {
   return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen name="Home" component={Homepage} />
-        <Stack.Screen name="Profile" component={Profile} />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <View>
+      <Homepage />
+      <NavigationBar />
+    </View>
   );
-};
+}
 
 const styles = StyleSheet.create({
   container: {
