+++ components/ProfilePicture.tsx
@@ -1,12 +1,13 @@
-import react from 'react'
+import react from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 
-export default function ProfilePicture () {
+export default function ProfilePicture() {
   return (
-    <div style={{backgroundColor: "#000000", color: "#FFFFFF", padding: "10px"}}>
-        <p>live laugh cope</p>
-        <p>i need a new major</p>
-        <p>(pfp under construction)</p>
-    </div>
+    <View style={{ backgroundColor: '#000000', padding: '10px' }}>
+      <Text style={{ color: '#FFFFFF' }}>live laugh cope</Text>
+      <p></p>
+      <p>i need a new major</p>
+      <p>(pfp under construction)</p>
+    </View>
   );
 }
