+++ screens/Profile.tsx
@@ -1,7 +1,7 @@
 import { StyleSheet, Text, View } from 'react-native';
 import ProfilePicture from '../components/ProfilePicture';
 
-export default function Profile() {
+export const Profile = () => {
   return (
     <View style={styles.container}>
       <Text style={styles.row}>Edit Profile</Text>
@@ -12,7 +12,7 @@ export default function Profile() {
       <Text style={styles.row}>Uhhhh pfp data goes here :(</Text>
     </View>
   );
-}
+};
 
 const styles = StyleSheet.create({
   container: {
