--- screens/auth/Signup.css
@@ -1,22 +1,0 @@
-.box {
-    font-size: medium;
-    width: 80%; 
-    height: 80%; 
-    padding: 10px; 
-    border-color: #f2f2f2; 
-    background-color: #f2f2f2;
-}
-
-.label {
-    font-weight: bold; 
-    font-size: medium; 
-    display: block; 
-    margin-bottom: 10px;
-    margin-top: 30px; 
-}
-
-.btn {
-    text-align: center;
-    width: 90%; 
-    margin-bottom: 50px; 
-}
